Adds reading acitivity views
