+++ bookwyrm/views/__init__.py
@@ -10,6 +10,8 @@ from .interaction import Favorite, Unfavorite, Boost, Unboost
 from .invite import ManageInvites, Invite
 from .landing import About, Home, Feed, Discover
 from .notifications import Notifications
+from .reading import edit_readthrough, create_readthrough, delete_readthrough
+from .reading import start_reading, finish_reading
 from .password import PasswordResetRequest, PasswordReset, ChangePassword
 from .tag import Tag, AddTag, RemoveTag
 from .search import Search
